Add infra build script to IPA
We now do builds of the IPA CoreOS ramdisk in the post script. (Ref:
openstack-infra/config/tree/modules/openstack_project/files/
jenkins_job_builder/config/ironic-python-agent-jobs.yaml)
We should store the build script for this in the local repo so we can
modify it as needed without needing commits to openstack-infra/config.

Change-Id: I92655c0552179613c23b129d1e2dce2ef1643391