Migrate CoreOS to use coreos-cloudinit
http://coreos.com/docs/cluster-management/setup/cloudinit-cloud-config/
describes how to use coreos-cloudinit to configure a node. Migrate to
using this instead of a single BASH script (which is deprecated).

This also means we need the latest version of CoreOS, so I've updated
coreos-oem-inject.py to pull from the most recent copies, and this
includes moving to the amd64-usr builds.

Change-Id: I620b8fec1bfc5d58b560434d17b4f617734fadb8